Attempt to fix threeblocking test
Give backup_B and backup_C enough time to reach the Plugin code. If A
gets the lock before backup_C even reached the Threadcounter, it will
get executed only after everything else is finished and lead to a
different test result.
This may be what is going on e.g. here:
https://travis-ci.org/signal-bot/signal-bot/builds/399336771